changed provisioned throughput for DynamoDB indexes
A default throughput of 5 read / 5 writes for each index should be really more than enough for most users.

Objects saved in DDB are really not so big, and most users will not make many builds and releases per second. It shouldn't need to be 80 for the total (about $50/month), 5 units each max (it's in operations / second, depending on the size of the objects retrieved) should largely suffice, and those who are consistantly using more than that will probably be able to figure out that they are really doing so many builds and releases.

FYI, I have set my company's at 3 units each, and it is so low I can barely see that data in the cloudwatch alarms that will usually show "insufficient data" most of the time at about 50 builds/day, even when I tried hard to build and release a bunch of apps at about the same time.