Ensure the correct environment when `convox run` is used with a specific release
If no releaseId is provided, `convox run` will use the last promoted release to run a
process. Otherwise iterate over the previous revisions (starting with the
latest) looking for the releaseId specified. This makes sure the same
environment variables are available.

This does not resolve the issue where `convox run` can be used with a
release that hasn't been promoted yet. That will be tackled separately.

Closes #669