Listing processes use a cache to prevent throttling
When gathering information for a large number of processes (e.g > 300), API throttling becomes an issue. This is addressed by caching various resources that are unlikely to change while gathering the process data. That includes container definitions and instances that are shared amongst the processes. I was able to gather over ~420 processes without being throttled back to back. Fixes #1990

Other fixes include:
- Bumping CircleCI golang to 1.7.5 as rack itself
- Added the `ProcessGet` interface so that `convox ps info <ID>` works. Fixes #2021
- Added some logging to the `SystemGet` interface