No blocking http requests in router
While inspecting why the local k8s rack on gen2 is still unusable for us I might have a new theory.

It seems like when our next.js dev server is compiling, which sometimes can take up to a minute, it blocks the router. This in turn makes all the healthchecks fail in kubernetes which causes everything to restart and behave generally bad. Next request is the same as the cache is now cleared because of the restart etc...

Is it possible that this PR could fix it?