[#3352] Rack param: ImagePullBehavior
Allow racks to configure ECS_IMAGE_PULL_BEHAVIOR for the ECS agent. https://docs.aws.amazon.com/AmazonECS/latest/developerguide/ecs-agent-config.html

I can't think of a reason not to do this. The default behavior is to pull the image every time but since image tags are NOT reused in Convox, it's safe to use the cache. Leaving the cache behavior as "opt-in" so there are no surprises but I'd encourage folks with large images to make the switch.