[#3413] Pull custom Rack tags down to the App stack
Users can apply their own custom tags to the Rack CloudFormation stack. That will cascade to any resources from that but not the Apps themselves. This change means an App release will pull in any custom Rack tags, which will then cascade to any App resources and the service and tasks in ECS.