[#3563] Using LogDriver to send logs to CloudWatch, Syslog or disable it
By default, CloudWatch is enabled, it can be changed to `Syslog` or turned off setting `LogDriver` empty.

`SyslogDestination` must be provided if `LogDriver` is `Syslog`.

```
 $ convox rack params set LogDriver=Syslog
ERROR: to enable Syslog as the LogDriver you must pass SyslogDestination parameter
```

Syslog format can be changed on`SyslogFormat`, the default is `rfc5424` (same as the current Syslog resource).