[#3543] Classic to Application Load Balancer migration
Implementation for migrating the rack API Classic Load Balancer to an Application Load Balancer.

- cloudformation changes
- new cli command `convox rack sync` to retrieve the new rack API URL

It's important to note that with the ELB migration, the rack API hostname will change.
The switch between hostname will be transparent for console managed racks. For users running racks outside of the console, we provide the `convox rack sync --name rack-name` command that should be executed after the update. This will print the new hostname to stdout.