[#3613] EFS resource for volumes
### What is the feature/fix?

A new way to have a persistent volume can be accessed between services (across instances and AZs). The EFS resource will allocate a new EFS volume that can be linked to services and used in the volumes.

### Add screenshot or video (optional)

```
resources:
  sharedvolume:
    type: efs
    options:
      path: "/bitnami"

environment:
  - PORT=3000
  - ENVIRONMENT=master
services:
  web:
    build: .
    port: 3000
    volumes:
      - /my/shared/data
      - /var/www/html
      - /sys/fs/cgroup/:/host/sys/fs/cgroup/
      - /proc/:/host/proc/
      - /var/run/docker.sock:/var/run/docker.sock
      - testvolume:/app
      - sharedvolume:/bitnami
    resources:
      - sharedvolume
```

### Does it has a breaking change?

No.

### How to use/test it?

Install a rack with the RC version (to be created, declare the EFS resource, link the resource in the service and use it in the volumes. You can see a snippet above.

### Checklist
- [ ] New coverage tests
- [ ] Unit tests passing
- [ ] E2E tests passing
- [ ] E2E downgrade/update test passing
- [ ] Documentation updated
- [ ] No warnings or errors on Deepsource/Codecov