[#3624] Worker monitor should not get LogGroup from Rack API service
### What is the feature/fix?

Since the Monitor and Web API had their name changed to include in the beginning, the worker was mistakenly assuming that the stack was from an app.

The monitor worker should only get the LogGroup from service apps.

### Does it has a breaking change?

No.

### How to use/test it?

Install the RC and run `convox rack logs`, it shouldn't show any error.

### Checklist
- [ ] New coverage tests
- [x] Unit tests passing
- [ ] E2E tests passing
- [ ] E2E downgrade/update test passing
- [ ] Documentation updated
- [x] No warnings or errors on Deepsource/Codecov