[#3628] Private subnets should only be created when rack is Private
### What is the feature/fix?

When the VPC endpoints and the private subnets were removed here https://github.com/convox/rack/pull/3607, the Timer was still using the SubnetPrivateX in the VPC config, which was blocking any attempt on updating any rack with Timers.

![image](https://user-images.githubusercontent.com/8239709/211540945-4ed809cb-3c68-44d6-a1cc-51e2b39a60de.png)

Since the Timers don't use the private subnet anymore, is safe to remove them if the rack is not private.

### Does it has a breaking change?

For old racks before `20221202145130` with Timers deployed, they will receive an error while updating the rack (due to the removal of the private subnet). The user must update to `20221202145130`, deploy the Timer again, and then update the rack.

### How to use/test it?

Update any rack to the RC, after the update is complete, check on VPC if the private subnet was deleted. For private racks, the private subnet should still exist.

### Checklist
- [ ] New coverage tests
- [x] Unit tests passing
- [x] E2E tests passing
- [ ] E2E downgrade/update test passing
- [ ] Documentation updated
- [x] No warnings or errors on Deepsource/Codecov