[#3631] Save racks installed locally through the CLI
### What is the feature/fix?

When a rack is installed using the CLI locally, not through the console or local rack, the installation finishes but don't keep the rack host anywhere. The CLI now will store the racks installed through it in the `self-managed` file.

### Does it has a breaking change?

No, however past racks installed using the CLI won't be recoverable, only if the user creates the rack entry on `self-managed`.

### How to use/test it?

1. Login into a console (e.g. console.convox.com)
2. Install the rack through the RC CLI (to be created).
3. Check if the rack is listed on `convox racks`
4. Switch to the new rack
5. Uninstall it
6. Run `convox racks` and the rack should not be on the list

### Checklist
- [x] New coverage tests
- [x] Unit tests passing
- [ ] E2E tests passing
- [ ] E2E downgrade/update test passing
- [ ] Documentation updated
- [ ] No warnings or errors on Deepsource/Codecov