[#3654] Use SSM Parameter store to keep the API Password on task definition
### What is the feature/fix?

Instead of passing the API password to the rack as plain text, it will store the password on SSM Parameter Store and then pass it as a secret to the task definition.

AWS charges to keep and use the Secret Manager, it would increase the user cost to store the password, the SSM is a cheaper alternative.

### Does it has a breaking change?

No

### How to use/test it?

Update/create a rack using the RC, then use it normally, it should not interfere in the user experience.

### Checklist
- [ ] New coverage tests
- [x] Unit tests passing
- [x] E2E tests passing
- [x] E2E downgrade/update test passing
- [ ] Documentation updated
- [x] No warnings or errors on Deepsource/Codecov