[#3656] Restricted the ApiRole permissions
### What is the feature/fix?

Restricted the ApiRole permissions: reduced the access level of the IAM role and policy associated with Convox's API service, to only allow the necessary actions to be performed.

### Does it has a breaking change?

No.

### How to use/test it?

Install/update a rack to the RC version and use/test an app, you may use the app inside of `examples/httpd`.

### Checklist
- [ ] New coverage tests
- [ ] Unit tests passing
- [ ] E2E tests passing
- [ ] E2E downgrade/update test passing
- [ ] Documentation updated
- [ ] No warnings or errors on Deepsource/Codecov