Better document a bad SHA256 invocation
This code uses sha256.New().Sum instead of sha256.Sum256. See
https://go.dev/play/p/vSW0U3Hq4qk for a demonstration of the differences.

SHA256 is used to generate identifiers that map server names to settings. I
think these IDs have been persisted to external sources (setting.go mentions S3
buckets), so moving this to a good invocation is hard and probably not worth
it. Instead, document the exact behavior of what's happening with the bad
invocation, make it more obvious, and enshrine it within a helper function.